Stripe.SubscriptionItem updated :quantity datatype
Resolves the issue https://github.com/code-corps/stripity_stripe/issues/626

+++ lib/stripe/subscriptions/subscription_item.ex
@@ -84,7 +84,7 @@ defmodule Stripe.SubscriptionItem do
                optional(:price) => Stripe.id() | Stripe.Price.t(),
                optional(:prorate) => boolean,
                optional(:proration_date) => Stripe.timestamp(),
-               optional(:quantity) => float,
+               optional(:quantity) => integer,
                optional(:tax_rates) => list(String.t())
              }
   def update(id, params, opts \\ []) do
